+++ semi/src/main/java/com/kh/semi/vo/memberCategoryVO.java
@@ -1,4 +1,4 @@
-package com.kh.semi.VO;
+package com.kh.semi.vo;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,8 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data @NoArgsConstructor @AllArgsConstructor @Builder
-public class memberRegionVO {
+public class memberCategoryVO {
 	String memberId;
-	int regionNo;
-	String regionType; //집/직장/관심지역
+	int categoryNo;
 }
