+++ semi/src/main/java/com/kh/semi/vo/memberRegionVO.java
@@ -1,4 +1,4 @@
-package com.kh.semi.VO;
+package com.kh.semi.vo;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,7 +6,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data @NoArgsConstructor @AllArgsConstructor @Builder
-public class memberCategoryVO {
+public class memberRegionVO {
 	String memberId;
-	int categoryNo;
+	int regionNo;
+	String regionType; //집/직장/관심지역
 }
